refactor: i32 -> i64

+++ crates/api_models/src/refunds.rs
@@ -8,7 +8,7 @@ pub struct RefundRequest {
     pub refund_id: Option<String>,
     pub payment_id: String,
     pub merchant_id: Option<String>,
-    pub amount: Option<i32>,
+    pub amount: Option<i64>,
     pub reason: Option<String>,
     //FIXME: Make it refund_type instant or scheduled refund
     pub force_process: Option<bool>,
@@ -19,7 +19,7 @@ pub struct RefundRequest {
 pub struct RefundResponse {
     pub refund_id: String,
     pub payment_id: String,
-    pub amount: i32,
+    pub amount: i64,
     pub currency: String,
     pub reason: Option<String>,
     pub status: RefundStatus,
